Call rakismet's "spam?" method and train it with corrections
Rakismet used to have a controller mixin that would take care of setting
the spam status of the model, but that's been removed.  And we're
currently not taking advantage of akismet feedback mechanism when it
makes a mistake in classifying comments.

Call comment.spam? ourselves and assign to comment.akismet.  Also tell
Akismet about mistakes it made when a moderator flips the spam status of
a comment so it can learn from its mistakes.  Explicitly set the
Rakismet.request params on our model, because this doesn't happen
automatically either.  Add stubbing test coverage for these calls,
because it wasn't there before.

Remove the comment_start_as config flag, because we always scan
a new comment before it is saved.  It has no real meaning.