Avoid emailing about ActiveRecord::RecordNotFound (404)
I'm currently getting lots of emails due to bots visiting garbage
URLs on my site that don't exist.  They should (and do) return 404.
But I still get an email because the post finder raises
ActiveRecord::RecordNotFound.

I'm not interested in getting emailed about this kind of error,
and generally returning a 404 is not considered an error.

Add ActiveRecord::RecordNotFound to the list of errors ignored by
ExceptionNotifier.